completed auth flow

+++ src/main/java/com/bank/services/TransactionService.java
@@ -31,7 +31,7 @@ public class TransactionService {
 
         AccountEntity account = accountDAO.getAccountById(accountNumber);
         if (account == null) {
-            LogService.logintoDB((long)-1, LogDAO.Action.TRANSACTIONS,"Account is not available with bank","USER IP",LogDAO.Status.FAILURE);
+            LogService.logintoDB(-1, LogDAO.Action.TRANSACTIONS,"Account is not available with bank","USER IP",LogDAO.Status.FAILURE);
             throw new AccountNotFoundException(accountNumber);
         }
 
@@ -42,7 +42,7 @@ public class TransactionService {
         accountDTO.setBalance(newBalance);
         AccountEntity updatedAccount=AccountMapper.dtoToEntity(accountDTO);
         accountDAO.updateAccountDetails(updatedAccount);
-        LogService.logintoDB((long) user_id, LogDAO.Action.TRANSACTIONS,"Amount has been credited into User Account","USER IP",LogDAO.Status.SUCCESS);
+        LogService.logintoDB( user_id, LogDAO.Action.TRANSACTIONS,"Amount has been credited into User Account","USER IP",LogDAO.Status.SUCCESS);
 
         TransactionEntity transaction = new TransactionEntity();
         transaction.setFrom_account_id(null); // Since it's a deposit
@@ -60,7 +60,7 @@ public class TransactionService {
 
         AccountEntity accountEntity=accountDAO.getAccountById(accountNumber);
         if(accountEntity==null){
-            LogService.logintoDB((long)-1, LogDAO.Action.TRANSACTIONS,"Account is not available with bank","USER IP",LogDAO.Status.FAILURE);
+            LogService.logintoDB(-1, LogDAO.Action.TRANSACTIONS,"Account is not available with bank","USER IP",LogDAO.Status.FAILURE);
             throw new AccountNotFoundException(accountNumber);
         }
         AccountDTO accountDTO=AccountMapper.entityToDTO(accountEntity);
@@ -68,17 +68,17 @@ public class TransactionService {
         int user_id=accountDTO.getUserId();
 
         if(balance<debitAmount){
-            LogService.logintoDB((long)user_id, LogDAO.Action.TRANSACTIONS,"Sufficient balance is not available in the account","USER IP",LogDAO.Status.FAILURE);
+            LogService.logintoDB(user_id, LogDAO.Action.TRANSACTIONS,"Sufficient balance is not available in the account","USER IP",LogDAO.Status.FAILURE);
             throw new InsufficientFundsException(debitAmount+"Not available in the Account");
         }
         else if(!accountDTO.isLocked()){
             accountDTO.setBalance(balance-debitAmount);
             accountEntity=AccountMapper.dtoToEntity(accountDTO);
             accountDAO.updateAccountDetails(accountEntity);
-            LogService.logintoDB((long)user_id, LogDAO.Action.TRANSACTIONS,"Amount has been debited from user account","USER IP",LogDAO.Status.SUCCESS);
+            LogService.logintoDB(user_id, LogDAO.Action.TRANSACTIONS,"Amount has been debited from user account","USER IP",LogDAO.Status.SUCCESS);
         }
         else{
-            LogService.logintoDB((long)user_id, LogDAO.Action.TRANSACTIONS,"Cannot debit from FD(Fixed Deposit)","USER IP",LogDAO.Status.FAILURE);
+            LogService.logintoDB(user_id, LogDAO.Action.TRANSACTIONS,"Cannot debit from FD(Fixed Deposit)","USER IP",LogDAO.Status.FAILURE);
             throw new BankingException("Cannot debit from FD(Fixed Deposit)");
         }
